requirements.txt
Django==1.8.4
Markdown==2.6.2
PyJWT==1.4.0
argparse==1.2.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.8.1
django-ckeditor==5.0.2
django-cors-headers==1.1.0
django-filter==0.11.0
django-guardian==1.3
django-import-export==0.2.8
django-oauth-toolkit==0.9.0
django-registration==1.0
django-rest-framework-social-oauth2==1.0.2
django-toolbelt==0.0.1
django-wysiwyg==0.7.1
djangorestframework==3.2.4
djoser==0.4.0
gunicorn==19.1.1
oauthlib==1.0.1
psycopg2==2.5.1
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static==0.4
static3==0.6.1
tablib==0.10.0
wsgiref==0.1.2
